PR-E-12: thin ReflectionMetadataEmitter root (FINAL Phase-2 PR)
Extracts the last two cohesive responsibility clusters from
ReflectionMetadataEmitter.cs and deduplicates accessibility mappers.
This is the final PR of the 22-PR decomposition; after this lands the
binder and emitter decomposition is complete.

New files
- CustomAttributeEncoder.cs (785 LoC) — owns every custom-attribute
  blob-encoding helper: EmitBoundAttribute, EmitUserAttributes,
  EmitStringAttribute, EmitIsReadOnlyAttributeOnParameter,
  ResolveAttributeConstructor, ParametersMatch, ArgAssignable,
  BuildCtorArgumentValues, NormalizeWellKnownType,
  IsTriviallyConvertible, EncodeClrTypeForCtorSig,
  WriteCustomAttributeFixedArg, WriteCustomAttributeArrayArg,
  GetSerializedTypeName, GetEnumUnderlyingTypeSafe,
  WriteCustomAttributeNamedArg, WriteCustomAttributeFieldOrPropertyType,
  and NextParameterHandle. Wired in EmitCore via a customAttrEncoder
  field initialised right after wellKnown; the assembly-level
  orchestrators (EmitReferenceAssemblyAttribute,
  EmitAssemblyInteropAttributes, EmitDebuggableAttribute,
  EmitNullableContextAttribute) stay on the root and forward into the
  encoder, per the plan.

- MethodBodyPlanner.cs (954 LoC) — the per-body planning surface that
  drives SlotPlanner's 16 collector classes. Owns CollectLocalsAndLabels,
  CollectStatements, CollectPatternSwitchSlots, the three Walk*
  forwarders, the thin CollectXxx() wrappers around the collector
  classes, CollectBlockExpressionLocals, CollectFunctionLiterals,
  CollectGoStatements, RegisterConstructedTypeAliases,
  CollectConstValues / WalkStmtsForConsts, the static PDB-shape
  helpers CollectLocalInfo and CollectLocalConstantInfo,
  AddIteratorInterfaceImplementations,
  AddAsyncIteratorInterfaceImplementations, GetSmPackage, and
  TryGetUserKickoffReceiverHandle. Constructed in EmitCore alongside
  the other components and back-bound to stateMachines via
  SetStateMachines once stateMachines exists.

  NeedsRvalueReceiverSpill / CanLoadVariableAddressForReceiverSpill /
  IsAddressableFieldAccessForReceiverSpill stay on the root because
  they form the delegate the root supplies to SlotPlanner's
  constructor — a bootstrap cycle prevents lifting them.

- AccessibilityMap.cs (64 LoC) — single canonical home for the four
  accessibility mappers (MapTypeAccessibility,
  MapNestedTypeAccessibility, MapFieldAccessibility,
  ToMethodVisibility). PR-E-8 had noted that TypeDefEmitter shipped its
  own copies; this PR de-duplicates both sites onto
  AccessibilityMap.X(...).

- MethodInfoHelpers.cs (173 LoC) — small static cluster of the
  method-vs-interface comparison helpers: RequiresVirtualOnValueType
  (#409), MethodImplicitlyImplementsInterface (handles both G#
  interfaces and imported CLR interfaces, #525), MethodSignaturesMatch,
  and CallableParameters.

Modified files
- ReflectionMetadataEmitter.cs: 6,149 LoC -> 4,399 LoC (-1,750, -28.5 %).
  Cluster A, Cluster B, AccessibilityMap, and MethodInfoHelpers
  bodies removed; all call sites rewired to the new homes.
- TypeDefEmitter.cs: accessibility-mapper copies removed; call sites
  rewired to AccessibilityMap.X(...). Doc updated.

Stretch target
- The plan's stretch target was < 1,500 LoC on
  ReflectionMetadataEmitter.cs. The PR did not hit that ceiling.
  Final root size is 4,399 LoC. The remaining bulk is split roughly
  between (a) the EmitFunction orchestrator (~390 LoC, kept per plan),
  (b) EmitGlobalFieldDefs (~30 LoC, kept per plan), (c) the five
  body-bytes shims that bridge TypeDefEmitter / StateMachineEmitter to
  MethodBodyEmitter (~550 LoC; the plan explicitly lists these as
  staying on the root because they are the seam between byte-needing
  callers and the body emitter), (d) the metadata-token resolvers and
  signature/CLR-type encoders (~1,100 LoC) and the assembly-level
  Emit*Attribute orchestrators (kept per plan). Lifting any of these
  further would require touching IL-emitting code; the plan's stop
  condition ("if hitting the LoC target requires forcing extractions
  that perturb IL, prefer to merge with a larger root + documented
  leftover than to introduce IL drift") applies. Documented leftover.

Acceptance
- RefactoringBaselineTests: zero hash diffs (the IL byte-identity
  gate) after each checkpoint.
- DeterministicEmitTests: green.
- SlotDictionaryAliasingAssertionTests: green (the substrings stay in
  the globbed Emit/*.cs set).
- Core.Tests: 2,120 passed, 0 failed, 1 skipped
  (RegenerateBaseline, as designed).
- Public/internal API surface visible to Compilation.cs unchanged:
  ReflectionMetadataEmitter keeps its ctor, Emit entry point, and
  public AssemblyName / AssemblyVersion / MetadataOnly /
  DebugInformation properties.
- All GS#### diagnostic codes / locations / messages unchanged.

Phase 2 + entire decomposition: complete.
- Phase 1 (Binder): 18,548 -> ~2,766 LoC across 9 components, 10 PRs
  (PR-0 and PR-B-1..PR-B-9).
- Phase 2 (Emitter): 18,016 -> 4,399 LoC across 12 components, 12 PRs
  (PR-E-1..PR-E-12).
- 22 PRs total, zero IL hash diffs at every merge.
