+++ src/Core/CodeAnalysis/Emit/MethodInfoHelpers.cs
@@ -1,0 +1,173 @@
+// <copyright file="MethodInfoHelpers.cs" company="GSharp">
+// Copyright (C) GSharp Authors. All rights reserved.
+// </copyright>
+
+using System.Collections.Immutable;
+using GSharp.Core.CodeAnalysis.Symbols;
+
+namespace GSharp.Core.CodeAnalysis.Emit;
+
+/// <summary>
+/// PR-E-12: small grouping of static method-info utilities lifted from
+/// <see cref="ReflectionMetadataEmitter"/>. These four helpers form a
+/// tight cluster around method-vs-interface comparison and signature
+/// matching:
+/// <list type="bullet">
+///   <item><description><see cref="RequiresVirtualOnValueType"/> — issue #409, the question
+///     "must this struct method keep <c>virtual</c> attributes because the CLR will
+///     vtable-dispatch through it?"</description></item>
+///   <item><description><see cref="MethodImplicitlyImplementsInterface"/> — supports the answer by
+///     walking both G# <see cref="StructSymbol.Interfaces"/> and imported CLR
+///     <see cref="StructSymbol.ImplementedClrInterfaces"/> (issue #525).</description></item>
+///   <item><description><see cref="MethodSignaturesMatch"/> — name/return/parameter equality
+///     between two G# functions, used when matching candidate implementations
+///     to interface members.</description></item>
+///   <item><description><see cref="CallableParameters"/> — strips the explicit receiver from a
+///     method's parameter list to get the "as-called" arity.</description></item>
+/// </list>
+/// Kept static and pure: no <see cref="EmitContext"/> reference required.
+/// </summary>
+internal static class MethodInfoHelpers
+{
+    /// <summary>
+    /// Issue #409: determines whether a value-type instance method must keep
+    /// virtual method attributes because it participates in CLR vtable dispatch.
+    /// </summary>
+    /// <param name="function">The function being inspected.</param>
+    /// <param name="receiverStruct">The struct that declares <paramref name="function"/>.</param>
+    /// <returns>True if the method must be emitted as <c>virtual</c>.</returns>
+    public static bool RequiresVirtualOnValueType(FunctionSymbol function, StructSymbol receiverStruct)
+    {
+        if (function.IsOverride || function.OverriddenMethod != null)
+        {
+            return true;
+        }
+
+        return MethodImplicitlyImplementsInterface(receiverStruct, function);
+    }
+
+    /// <summary>
+    /// Determines whether a method on a class/struct implicitly implements an
+    /// interface method (same name, parameters, and return type). Considers
+    /// both G# interfaces (<see cref="StructSymbol.Interfaces"/>) and imported
+    /// CLR interfaces (<see cref="StructSymbol.ImplementedClrInterfaces"/>,
+    /// issue #525).
+    /// </summary>
+    /// <param name="structSym">The struct that contains <paramref name="method"/>.</param>
+    /// <param name="method">The candidate method.</param>
+    /// <returns>True if the method matches any interface method (G# or CLR).</returns>
+    public static bool MethodImplicitlyImplementsInterface(StructSymbol structSym, FunctionSymbol method)
+    {
+        if (!structSym.Interfaces.IsDefaultOrEmpty)
+        {
+            foreach (var iface in structSym.Interfaces)
+            {
+                if (iface.Methods.IsDefaultOrEmpty)
+                {
+                    continue;
+                }
+
+                foreach (var ifaceMethod in iface.Methods)
+                {
+                    if (MethodSignaturesMatch(ifaceMethod, method))
+                    {
+                        return true;
+                    }
+                }
+            }
+        }
+
+        if (!structSym.ImplementedClrInterfaces.IsDefaultOrEmpty)
+        {
+            var callable = CallableParameters(method);
+            var returnClr = method.Type?.ClrType;
+            foreach (var ifaceSym in structSym.ImplementedClrInterfaces)
+            {
+                var clrIface = ifaceSym?.ClrType;
+                if (clrIface == null)
+                {
+                    continue;
+                }
+
+                foreach (var clrMethod in clrIface.GetMethods(System.Reflection.BindingFlags.Public | System.Reflection.BindingFlags.Instance))
+                {
+                    if (clrMethod.IsSpecialName || clrMethod.Name != method.Name)
+                    {
+                        continue;
+                    }
+
+                    var clrParams = clrMethod.GetParameters();
+                    if (clrParams.Length != callable.Length)
+                    {
+                        continue;
+                    }
+
+                    if (returnClr != null && !ClrTypeUtilities.AreSame(returnClr, clrMethod.ReturnType))
+                    {
+                        continue;
+                    }
+
+                    var allMatch = true;
+                    for (var i = 0; i < clrParams.Length; i++)
+                    {
+                        var paramClr = callable[i].Type?.ClrType;
+                        if (paramClr == null || !ClrTypeUtilities.AreSame(paramClr, clrParams[i].ParameterType))
+                        {
+                            allMatch = false;
+                            break;
+                        }
+                    }
+
+                    if (allMatch)
+                    {
+                        return true;
+                    }
+                }
+            }
+        }
+
+        return false;
+    }
+
+    /// <summary>
+    /// Returns true when the two G# function signatures match (name, return
+    /// type, and ordered parameter types).
+    /// </summary>
+    /// <param name="interfaceMethod">The interface method.</param>
+    /// <param name="implementationMethod">The candidate implementation method.</param>
+    /// <returns>True if the signatures are equivalent.</returns>
+    public static bool MethodSignaturesMatch(FunctionSymbol interfaceMethod, FunctionSymbol implementationMethod)
+    {
+        if (interfaceMethod.Name != implementationMethod.Name || interfaceMethod.Type != implementationMethod.Type)
+        {
+            return false;
+        }
+
+        var interfaceParameters = CallableParameters(interfaceMethod);
+        var implementationParameters = CallableParameters(implementationMethod);
+        if (interfaceParameters.Length != implementationParameters.Length)
+        {
+            return false;
+        }
+
+        for (var i = 0; i < interfaceParameters.Length; i++)
+        {
+            if (interfaceParameters[i].Type != implementationParameters[i].Type)
+            {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    /// <summary>
+    /// Returns the method's parameter list with the explicit receiver (the
+    /// first parameter when <see cref="FunctionSymbol.ExplicitReceiverParameter"/>
+    /// is set) removed — i.e. the parameters as seen by the call site.
+    /// </summary>
+    /// <param name="method">The method.</param>
+    /// <returns>The "as-called" parameter list.</returns>
+    public static ImmutableArray<ParameterSymbol> CallableParameters(FunctionSymbol method)
+        => method.ExplicitReceiverParameter == null ? method.Parameters : method.Parameters.RemoveAt(0);
+}
